+++ core/src/main/java/nl/gjorgdy/pl3xmarkers/core/interfaces/IStorage.java
@@ -2,7 +2,7 @@ package nl.gjorgdy.pl3xmarkers.core.interfaces;
 
 import nl.gjorgdy.pl3xmarkers.core.interfaces.entities.IAreaMarker;
 import nl.gjorgdy.pl3xmarkers.core.interfaces.entities.IIconMarker;
-import nl.gjorgdy.pl3xmarkers.core.interfaces.entities.ILineMarker;
+import nl.gjorgdy.pl3xmarkers.core.interfaces.entities.ISignMarker;
 
 public interface IStorage {
 
@@ -11,6 +11,8 @@ public interface IStorage {
 	IAreaMarkerRepository<? extends IAreaMarker> getAreaMarkerRepository();
 
 	IIconMarkerRepository<? extends IIconMarker> getIconMarkerRepository();
+
+	ISignMarkerRepository<? extends ISignMarker> getSignMarkerRepository();
 
 	void close();