+++ core/src/main/java/nl/gjorgdy/pl3xmarkers/core/interfaces/entities/IAreaMarker.java
@@ -1,5 +1,7 @@
 package nl.gjorgdy.pl3xmarkers.core.interfaces.entities;
 
+import org.jspecify.annotations.Nullable;
+
 import java.util.Collection;
 import java.util.Locale;
 
@@ -12,6 +14,12 @@ public interface IAreaMarker {
 	int getColor();
 
 	Collection<? extends IPoint> getPoints();
+
+	@Nullable
+	IPoint getMinCorner();
+
+	@Nullable
+	IPoint getMaxCorner();
 
 	boolean addPoint(int x, int z);