+++ core/src/main/java/nl/gjorgdy/pl3xmarkers/core/interfaces/entities/IIconMarker.java
@@ -8,4 +8,8 @@ public interface IIconMarker {
 
 	IPoint getLocation();
 
+	default String getKey() {
+		return getLocation().getX() + ":" + getLocation().getZ();
+	}
+
 }