+++ json/src/main/java/nl/gjorgdy/pl3xmarkers/core/json/JsonStorage.java
@@ -6,6 +6,7 @@ import nl.gjorgdy.pl3xmarkers.core.interfaces.IStorage;
 import nl.gjorgdy.pl3xmarkers.core.interfaces.entities.ILineMarker;
 import nl.gjorgdy.pl3xmarkers.core.json.repositories.AreaMarkerRepository;
 import nl.gjorgdy.pl3xmarkers.core.json.repositories.IconMarkerRepository;
+import nl.gjorgdy.pl3xmarkers.core.json.repositories.SignMarkerRepository;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -13,7 +14,8 @@ import java.nio.file.Path;
 
 public class JsonStorage implements IStorage {
 
-	private  IconMarkerRepository iconMarkerRepository;
+	private IconMarkerRepository iconMarkerRepository;
+	private SignMarkerRepository signMarkerRepository;
 	private AreaMarkerRepository areaMarkerRepository;
 	private boolean loaded = false;
 
@@ -21,9 +23,12 @@ public class JsonStorage implements IStorage {
 		// read files
 		String configPath = Pl3xMarkersCore.isBukkit() ? "plugins/Pl3xMarkers" : "config/pl3xmarkers";
 		// migrate old data if needed
-		if (Pl3xMarkersCore.isBukkit()) migrate(Path.of("config/pl3xmarkers"), Path.of("plugins/Pl3xMarkers"));
+		if (Pl3xMarkersCore.isBukkit()) {
+			migrate(Path.of("config/pl3xmarkers"), Path.of("plugins/Pl3xMarkers"));
+		}
 		// load repositories
 		iconMarkerRepository = new IconMarkerRepository(configPath + "/markers", "icons");
+		signMarkerRepository = new SignMarkerRepository(configPath + "/markers", "signs");
 		areaMarkerRepository = new AreaMarkerRepository(configPath + "/markers", "areas");
 		// mark as loaded
 		loaded = true;
@@ -36,14 +41,26 @@ public class JsonStorage implements IStorage {
 
 	@Override
 	public AreaMarkerRepository getAreaMarkerRepository() {
-		if (!loaded) load();
+		if (!loaded) {
+			load();
+		}
 		return areaMarkerRepository;
 	}
 
 	@Override
 	public IconMarkerRepository getIconMarkerRepository() {
-		if (!loaded) load();
+		if (!loaded) {
+			load();
+		}
 		return iconMarkerRepository;
+	}
+
+	@Override
+	public SignMarkerRepository getSignMarkerRepository() {
+		if (!loaded) {
+			load();
+		}
+		return signMarkerRepository;
 	}
 
 	@Override
@@ -56,9 +73,12 @@ public class JsonStorage implements IStorage {
 	}
 
 	private void writeInternal() {
-		if (!loaded) load();
+		if (!loaded) {
+			load();
+		}
 		// save files
 		iconMarkerRepository.write();
+		signMarkerRepository.write();
 		areaMarkerRepository.write();
 	}
 
@@ -70,9 +90,11 @@ public class JsonStorage implements IStorage {
 			files.forEach(oldFile -> {
 				try {
 					Files.move(oldFile, newPath.resolve(oldFile.getFileName()));
-				} catch (Exception ignored) {}
+				} catch (Exception ignored) {
+				}
 			});
 			Files.delete(oldPath);
-		} catch (IOException ignored) {}
+		} catch (IOException ignored) {
+		}
 	}
 }